+++ test/test_custom_fields_discover.py
@@ -1,10 +1,11 @@
-from tap_tester import menagerie, connections, runner
-
 from base import ZendeskTest
+from tap_tester import connections, menagerie, runner
+
 
 class ZendeskCustomFieldsDiscover(ZendeskTest):
     def name(self):
         return "tap_tester_zendesk_custom_fields_discover"
+
 
     def test_run(self):
         # Default test setup
@@ -40,4 +41,3 @@ class ZendeskCustomFieldsDiscover(ZendeskTest):
             # with their own set of properties
             self.assertIsNotNone(properties.get('{}_fields'.format(schema[0]), {}).get('properties'),
                                  msg='{}_fields not present in schema!'.format(schema[0]))
-