+++ test/test_pagination.py
@@ -1,7 +1,5 @@
-import tap_tester.connections as connections
-import tap_tester.runner as runner
-import tap_tester.menagerie as menagerie
 from base import ZendeskTest
+from tap_tester import connections, menagerie, runner
 
 
 class ZendeskPagination(ZendeskTest):
@@ -12,21 +10,23 @@ class ZendeskPagination(ZendeskTest):
     def name(self):
         return "zendesk_pagination_test"
 
+
     def test_run(self):
         """
-        • Verify that for each stream you can get multiple pages of data.  
+        • Verify that for each stream you can get multiple pages of data.
         This requires we ensure more than 1 page of data exists at all times for any given stream.
         • Verify by pks that the data replicated matches the data we expect.
 
         Outstanding Work:
         TDL-17980 [tap-zendesk][tap-tester] Enable CRUD operations for `tags` stream to stabilize pagination test
         """
-        
+
         # Streams to verify all fields tests
         expected_streams = self.expected_check_streams()
         expected_streams = expected_streams - {
             "satisfaction_ratings", # skip as only end user of tickets can create data
             "tags", #  Test Stability Issue: TDL-17980
+            "talk_phone_numbers"
         }
 
         conn_id = connections.ensure_connection(self)
@@ -53,7 +53,7 @@ class ZendeskPagination(ZendeskTest):
 
                 # expected values
                 expected_primary_keys = self.expected_primary_keys()[stream]
-                
+
                 # verify that we can paginate with all fields selected
                 record_count_sync = record_count_by_stream.get(stream, 0)
                 self.assertGreater(record_count_sync, self.API_LIMIT,