+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # Changelog
 
-## 2.0.1
+## 2.3.1
   * Dependabot update [#129](https://github.com/singer-io/tap-zendesk/pull/129)
+## 2.3.0
+  * Adds configurable page size for requests [#141](https://github.com/singer-io/tap-zendesk/pull/141)
+## 2.2.0
+  * Adds Support for lookup fields [#124](https://github.com/singer-io/tap-zendesk/pull/124)
 
+## 2.1.0
+  * Adds new streams `talk_phone_numbers` and `ticket_metric_events` [#111](https://github.com/singer-io/tap-zendesk/pull/111)
+## 2.0.1
+  * Adds backoff/retry for `ProtocolError` and `ChunkedEncodingError` [#131](https://github.com/singer-io/tap-zendesk/pull/131)
+## 2.0.0
+  * Incremental Exports API implementation for User's stream [#127](https://github.com/singer-io/tap-zendesk/pull/127)
+## 1.7.6
+  * Fix Infinite Loop for Users [#103](https://github.com/singer-io/tap-zendesk/pull/103)
 ## 1.7.5
   * Added support for backoff and retry for error 409 [#107](https://github.com/singer-io/tap-zendesk/pull/107)
   * Code Formatting [#107](https://github.com/singer-io/tap-zendesk/pull/107)